to float32

+++ globalpricematrix.py
@@ -41,7 +41,7 @@ class GlobalPriceMatrix(CoinList):
 	coin = COIN_REF
 	chart = self.getChart(coin, start = self._start, end = self._end)
 	cols = [d['date'] for d in chart]
-	self._pm = pd.DataFrame(index = self._coins, columns = cols)
+	self._pm = pd.DataFrame(index = self._coins, columns = cols).astype('float32')
 	self.__fillPriceRow(coin, start = self._start, end = self._end)
 
 	for c in self._coins:
@@ -84,7 +84,7 @@ class GlobalPriceMatrix(CoinList):
 
 
     def __getPriceFromFile(self, csv = CSV_DEFAULT):
-	pm = pd.DataFrame.from_csv(csv)
+	pm = pd.DataFrame.from_csv(csv).astype('float32')
 	pm.index = pm.index.astype(np.int64)/10**9
 	self._pm = pm.transpose()
 	self._start = self._pm.columns[0]
